Fix Proxy Example
The proxy class should be constructed the same way as the class it wraps. From this SO answer: "Proxy and Decorator both have the same interface as their wrapped types, but the proxy creates an instance under the hood, whereas the decorator takes an instance in the constructor." https://stackoverflow.com/a/7211706/2433572

As it stands, the proxy pattern is indistinguishable from the decorator pattern.