Apply SOLID principle at Door code example
Many examples on the internet of software engineering didn't apply
the SOLID principle correctly. Considering the Doors could be made
in many kinds of materials, it sounds more like an Abstract class.

The changes there introduced will be explained over the SOLID
principle:
* A Measurable Interface is a behavior expected for all objects which
have dimensions, fitting to the Interface Segregation Principle.
* Abstract Door has dimensions and requires a material to be built.
It is a requirement for the Single Responsibility Principle.
* Wooden Door is the concrete implementation over Door abstraction
where defines its material and dimensions. This is an example of the
Open/Close Principle.